+++ spec/superset/dataset/warm_up_cache_spec.rb
@@ -1,56 +1,19 @@
 require 'spec_helper'
 
 RSpec.describe Superset::Dataset::WarmUpCache do
-  subject { described_class.new(dashboard_id: dashboard_id) }
+  subject { described_class.new(dashboard_id: dashboard_id, table_name: table_name, db_name: db_name) }
   let(:dashboard_id) { 1 }
+  let(:table_name) { "Dataset 101"}
+  let(:db_name) { "Client Database 1" }
+  before do
+    allow(subject).to receive(:response).and_return(response)
+  end
 
   describe '.perform' do
-    context "Dataset count is not considered" do
-      let(:response) { nil }
-      before do
-        allow(subject).to receive(:response).and_return(response)
-      end
-      context 'when dashboard_id is not present' do
-        let(:dashboard_id) { nil }
+    let(:response) { 'Dataset warmed up' }
 
-        it 'raises an error' do
-          expect { subject.perform }.to raise_error(Superset::Request::InvalidParameterError, "dashboard_id must be present and must be an integer")
-        end
-      end
-
-      context 'when dashboard_id is not an integer' do
-        let(:dashboard_id) { 'string' }
-
-        it 'raises an error' do
-          expect { subject.perform }.to raise_error(Superset::Request::InvalidParameterError, "dashboard_id must be present and must be an integer")
-        end
-      end
-
-      context 'when dashboard_id is an integer' do
-        let(:response) { 'Dashboard warmed up' }
-
-        it 'warms up the dashboard' do
-          expect(subject.perform).to eq response
-        end
-      end
-    end
-
-    context 'when dashboard has multiple datasets' do
-      let(:dataset_details) do
-        [
-            {"name"=>"client database 1", "datasource_name"=>"datasource 101"},
-            {"name"=>"client database 2", "datasource_name"=>"datasource 102"},
-        ]
-      end
-      let(:api_response) { "Dataset warmed up" }
-      before do
-        allow(subject).to receive(:fetch_dataset_details).with(dashboard_id) { dataset_details } 
-        allow(subject).to receive(:api_response).and_return(api_response)
-      end
-      it 'warms up all the datasets' do
-        subject.response
-        expect(subject).to have_received(:api_response).exactly(dataset_details.count).times
-      end
+    it 'warms up the dataset' do
+      expect(subject.perform).to eq response
     end
   end
 end
